backend/core/TTS/requirements.txt
annotated-types
anyio
asttokens
boto3
botocore
certifi
charset-normalizer
colorama
decorator
elevenlabs
exceptiongroup
executing
h11
httpcore
httpx
idna
ipython
jedi
jmespath
matplotlib-inline
parso
prompt-toolkit
pure-eval
pydantic
pydantic_core
Pygments
python-dateutil
python-dotenv
requests
s3transfer
six
sniffio
stack-data
traitlets
typing_extensions
urllib3
wcwidth
websockets